Plane lab done
Fixed SOME issues with main game, cars still feel bad, ui is bugged, healthbars are bugged, take damage is bugged, objectives and scoring works, nothing to display score though. this is this project finished im done with it i have lost motivation and paitence for the time investment to fix this. no sound effects due to time limit as well.